Revert "Merge pull request #42 from project-safeway/feat/testes"
This reverts commit f8c677c1555c66228044634d3de280a3d1e26f09, reversing
changes made to bb2801085dd875fde01f182634ed87d43a1a594e.

+++ src/main/java/com/safeway/tech/services/AlunoService.java
@@ -37,7 +37,7 @@ public class AlunoService {
         aluno.setDtNascimento(request.dtNascimento());
         aluno.setSerie(request.serie());
         aluno.setSala(request.sala());
-        aluno.setValorMensalidade(BigDecimal.valueOf(request.valorMensalidade()));
+        aluno.setValorMensalidade(request.valorMensalidade());
         aluno.setDiaVencimento(request.diaVencimento());
         aluno.setAtivo(true);
